+++ src/parlant/core/engines/alpha/tool_caller.py
@@ -34,8 +34,15 @@ from parlant.core.nlp.generation_info import GenerationInfo
 from parlant.core.services.tools.service_registry import ServiceRegistry
 from parlant.core.sessions import Event, ToolResult
 from parlant.core.shots import Shot, ShotCollection
-from parlant.core.tools import Tool, ToolContext, ToolParameterDescriptor, ToolParameterOptions
-from parlant.core.tools import ToolId, ToolService
+from parlant.core.tools import (
+    Tool,
+    ToolContext,
+    ToolParameterDescriptor,
+    ToolParameterOptions,
+    ToolId,
+    ToolService,
+    DEFAULT_PARAMETER_PRECEDENCE,
+)
 
 ToolCallId = NewType("ToolCallId", str)
 ToolResultId = NewType("ToolResultId", str)
@@ -113,6 +120,7 @@ class MissingToolData:
     significance: Optional[str] = field(default=None)
     description: Optional[str] = field(default=None)
     examples: Optional[Sequence[str]] = field(default=None)
+    precedence: Optional[int] = field(default=DEFAULT_PARAMETER_PRECEDENCE)
 
 
 @dataclass(frozen=True)
@@ -302,7 +310,7 @@ class ToolCaller:
                             )
                             continue
 
-                        _, tool_options = tool.parameters[evaluation.parameter_name]
+                        tool_descriptor, tool_options = tool.parameters[evaluation.parameter_name]
 
                         if (
                             evaluation.is_missing
@@ -313,6 +321,8 @@ class ToolCaller:
                                 MissingToolData(
                                     parameter=evaluation.parameter_name,
                                     significance=tool_options.significance,
+                                    description=tool_descriptor.get("description"),
+                                    precedence=tool_options.precedence,
                                 )
                             )